+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased](https://github.com/dalance/nom-greedyerror/compare/v0.3.0...Unreleased) - ReleaseDate
 
+* [Added] ContextError implementation
+
 ## [v0.3.0](https://github.com/dalance/nom-greedyerror/compare/v0.2.0...v0.3.0) - 2020-11-26
 
 * [Added] nom 6.0.0 support